Move common files into the common folder
Files used by both the front and back end are moved into a new
folder called common.
