+++ arm_walker.h
@@ -20,7 +20,7 @@
 #include <stddef.h>
 
 #include "arm_core.h"
-#include "error.h"
+#include "common/error.h"
 
 typedef struct subtlis_arm_walker_t_ subtlis_arm_walker_t;
