+++ common/bitset.c
