+++ common/bitset.h
