+++ common/buffer.c
