+++ common/buffer.h
