+++ common/builtins.c
