+++ common/builtins.h
