+++ common/ir.c
