+++ common/string_pool.c
