+++ common/string_pool.h
