+++ common/type.c
