+++ expression.c
@@ -18,8 +18,8 @@
 #include <string.h>
 
 #include "builtins_ir.h"
+#include "common/ir.h"
 #include "expression.h"
-#include "ir.h"
 
 typedef int32_t (*subtilis_const_shift_t)(int32_t a, int32_t b);
