+++ unit_tests.c
@@ -19,7 +19,7 @@
 #include "arm_core_test.h"
 #include "arm_reg_alloc_test.h"
 #include "arm_test.h"
-#include "bitset_test.h"
+#include "common/bitset_test.h"
 #include "fpa_test.h"
 #include "ir_test.h"
 #include "lexer_test.h"
